separates embedded message out into view model
